Do not flush in the generate command. This is why:
If a sitemap providers needs to add thousands entries to the sitemap,
the developer will want to flush regularly to avoid having too much
documents waiting in the UnitOfWork.
For this reason I think it is better to let the developper call
`$sitemap->save`, instead of having the command only doing it at the
end. I updated the README to mention it.
Plus, it makes sitemap population coherent with data fixtures loading,
where the developer has the responsibility to flush the manager.